+++ SD.Web/SD.Services.Data/Services/Contracts/IUserSensorService.cs
@@ -14,7 +14,7 @@ namespace SD.Services.Data.Services.Contracts
 		Task<IPagedList<UserSensor>> GetSensorsByUserId(Guid userId, int pageNumber = 1, int pageSize = 10);
 
 		Task<UserSensor> AddUserSensorAsync(string userId, string sensorId, string name, string description, int latitude,
-			int longitude, int alarmMin, int alarmMax, int pollingInterval, bool alarmTriggered, bool isPublic);
+			int longitude, double alarmMin, double alarmMax, int pollingInterval, bool alarmTriggered, bool isPublic);
 
 		Task<UserSensor> GetSensorByIdAsync(string id);
