+++ SD.Web/SD.Web/Areas/Administration/Models/SensorViewModels/ModifyViewModel.cs
@@ -61,11 +61,11 @@ namespace SD.Web.Areas.Administration.Models.SensorViewModels
 
 		[Required]
 		[Display(Name = "AlarmMin")]
-		public int AlarmMin { get; set; }
+		public double AlarmMin { get; set; }
 
 		[Required]
 		[Display(Name = "AlarmMax")]
-		public int AlarmMax { get; set; }
+		public double AlarmMax { get; set; }
 
 		public IList<Tuple<string, string>> SensorNamesIds { get; set; }
 	}
